requirements.txt
Pillow
pyautogui
playsound
pytesseract
opencv-python
numpy 